backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "retriever-backend"
version = "0.1.0"
description = "Retriever Service MVP - Recall + Rerank"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0,<0.115.0",
    "uvicorn[standard]>=0.27.0,<0.32.0",
    "pymupdf>=1.24.0,<1.25.0",
    "python-pptx>=0.6.23,<0.7.0",
    "python-docx>=1.1.0,<1.2.0",
    "sqlalchemy>=2.0.25,<2.1.0",
    "pydantic>=2.6.0,<2.10.0",
    "pydantic-settings>=2.2.0,<2.6.0",
    "llama-index-core>=0.10.0,<0.11.0",
    "Pillow>=10.2.0,<11.0.0",
    "httpx>=0.27.0,<0.28.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]
